Tab: added update_inactive_cursors method
(update_inactive_cursors is still just a WIP)

We're now properly setting cursors of parent List_dir[s] BUT those (the
cursors) are set AFTER creating the columns themselves (therefore their
respective handler lambdas are called prior to calling
update_inactive_cursors), which causes those to be ignored.

In other news, update_cursor method is now renamed to
update_active_cursor (to make it obvious as to which cursor we're about
to update), get_active_column_num is renamed to (I hope) more sane name
get_current_ncols.

Lastly, get_list_dir_handler method was added - it is inlined because
the only 2 things it does are: checking the type of handler (and
throwing) and static_cast-ing to List_dir*.